Refactor project into bootstrap and hosted packages
