+++ lispfun/__init__.py
@@ -1,0 +1,11 @@
+"""LispFun package exposing the bootstrap interpreter by default."""
+from .bootstrap import (
+    parse, parse_multiple, to_string,
+    eval_lisp, standard_env,
+    Environment, Symbol, String, ListType,
+)
+__all__ = [
+    'parse', 'parse_multiple', 'to_string',
+    'eval_lisp', 'standard_env',
+    'Environment', 'Symbol', 'String', 'ListType'
+]
