+++ lispfun/bootstrap/AGENTS.md
@@ -1,0 +1,1 @@
+Refer to the repository root AGENTS.md for contribution guidelines.
