+++ lispfun/bootstrap/README.md
@@ -1,0 +1,6 @@
+# Bootstrap Interpreter
+
+This package implements the minimal Python interpreter used to bootstrap LispFun.
+It provides tokenising, parsing and evaluation in Python.  The interpreter is
+used by `run_bootstrap.py` and by the self-hosted evaluator when loading
+`eval2`.
