--- lispfun/evaluator.lisp
@@ -1,4 +1,0 @@
-(import "lispfun/list_utils.lisp")
-(import "lispfun/string_utils.lisp")
-(import "lispfun/eval_core.lisp")
-
