+++ lispfun/hosted/README.md
@@ -1,0 +1,5 @@
+# Self-hosted Evaluator
+
+This directory contains the Lisp source files for the `eval2` evaluator.  The
+Python runner `lispfun/run.py` loads `evaluator.lisp` from here and executes
+expressions via `(eval2 ...)`.
