+++ lispfun/hosted/__init__.py
@@ -1,0 +1,2 @@
+"""Lisp files for the self-hosted evaluator."""
+__all__ = []
