+++ lispfun/hosted/string_utils.lisp
