+++ toy/README.md
@@ -1,0 +1,4 @@
+# Toy Interpreter
+
+The `toy` directory contains a small Lisp interpreter implemented entirely in
+Lisp.  It is loaded by `run_toy.py` after the `eval2` evaluator is available.
